requirements.txt
Django==3.1.7
djangorestframework==3.12.2
mysqlclient==2.1.1
django-mysql==3.11.1
django-cors-headers==3.7.0
PyJWT==2.0.1
Faker==6.6.1
requests==2.30.0
pytest==7.3.1
pytest-django==4.5.2
django-allauth==0.54.0
dj-rest-auth==4.0.1
google-api-python-client==2.89.0



